Look up listing in DB upon GET, modify DB setup/seeding to accommodate
