+++ webpack.config.js
@@ -10,7 +10,7 @@ module.exports = {
   devtool: 'source-map',
   mode: 'development',
   resolve: {
-    extensions: ['*', '.js', '.jsx'],
+    extensions: ['*', '.js', '.jsx', '.json'],
   },
   module: {
     rules: [
